Create Point-Target-Response
This sets up targets, frequency signals, sets targets, and generates pulses to be sent to the targets and reflected back to be collected.  The reflection angle, intensity of the signal, positioning of the sensors,  are all important for creating  reconstructed image of the pulses traveling to the point and being reflected back.