While loop for Checking Server and Downloading DZT Files
The code runs through a while loop to check the URL contents on the
server to be observed.  The while loop is continuous through as many
repetitions as defined, and will display the links on the server every
10 seconds.  Once a new file has been uploaded to the server, the loop
can be stopped and the following code will download the new file
according to the inputted file name.